Re-factored the library to better separate basic utils from reproducibility tools.
