+++ cneuromod_embeddings/dypac_utils.py
@@ -30,7 +30,7 @@ def dypac_params():
 def key_params(atlas, fwhm, cluster, state):
     if (atlas=='intra') | (atlas=='inter'):
         params = f'{atlas}_fwhm-{fwhm}_cluster-{cluster}_state-{state}'
-    else: 
+    else:
         params = f'{atlas}_fwhm-{fwhm}'
     return params
 
@@ -42,30 +42,6 @@ def subject_keys(n_subject):
     return list_subject
 
 
-def visu_repro(root_data, xp, fwhm, cluster, state, n_subject=6):
-    R = np.load(os.path.join(root_data, xp, f'Rmatch_fwhm-{fwhm}_cluster-{cluster}_state-{state}.npy'))
-    width_fig = 20
-    n_comp = n_subject
-    fig = plt.figure(figsize=(width_fig, n_comp * 3))
-    for sub1 in range(n_subject):
-        for sub2 in range(n_subject):
-            plt.subplot(6, 6, 1 + sub1 + sub2 * 6)
-            match_val = np.max(R[:, :, sub1, sub2], axis=1)
-            plt.hist(match_val, bins=100, density=True, stacked=True)
-            plt.axis((-0.05, 1.05, 0, 15))
-            ax = plt.gca()
-            if sub2 == 0:
-                ax.set_title(f"match with sub-0{sub1+1}")
-            if sub1 != 0:
-                ax.tick_params(labelleft=False)
-            else:
-                ax.set_ylabel(f"sub-0{sub2+1} distribution")
-            if sub2 != 5:
-                ax.tick_params(labelbottom=False)
-            else:
-                ax.set_xlabel("spatial correlation")
-
-
 def normalize_components(model, mask_img):
     scaler = StandardScaler()
     x = model.components_.todense()
@@ -73,48 +49,6 @@ def normalize_components(model, mask_img):
     masker = NiftiMasker(standardize=False, detrend=False, mask_img=mask_img)
     xn = scaler.fit_transform(masker.fit_transform(img_parcels).transpose())
     return xn
-
-
-def R_models(model1, model2, mask_img):
-    xn = normalize_components(model1, mask_img)
-    xn = xn[:, model1.dwell_time_ > 0]
-    yn = normalize_components(model2, mask_img)
-    yn = yn[:, model2.dwell_time_ > 0]
-    R = np.matmul(xn.transpose(), yn) / yn.shape[0]
-    return R
-
-
-def match_components(sub1, sub2, root_data, fwhm, cluster, state):
-    model, mask_img1 = load_dypac(
-        subject=sub1,
-        root_data=root_data,
-        fwhm=fwhm,
-        cluster=cluster,
-        state=state,
-        batch="even",
-    )
-    if sub2 == sub1:
-        model2, mask_img2 = load_dypac(
-            subject=sub2,
-            root_data=root_data,
-            fwhm=fwhm,
-            cluster=cluster,
-            state=state,
-            batch="odd",
-        )
-    else:
-        model2, mask_img2 = load_dypac(
-            subject=sub2,
-            root_data=root_data,
-            fwhm=fwhm,
-            cluster=cluster,
-            state=state,
-            batch="even",
-        )
-    mask_img = math_img(
-        "img1 + img2 > 0", img1=model.masker_.mask_img, img2=model2.masker_.mask_img
-    )
-    return R_models(model, model2, mask_img)
 
 
 def load_model(pickle_in):
